Move `rename_function` to hy.macros
